pyproject.toml
[project]
name = "rorserverbot"
version = "2.0.0"
description = ""
authors = [
    {name = "Rafael Galvan",email = "rafael@zentro.codes"}
]
license = {text = "GNU GPLv3"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "aiohttp (>=3.13.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
